Use `Cow<'static, [u8]>` in image/svg, add constructors taking &[u8]`
This should resolve https://github.com/iced-rs/iced/issues/580 by
providing a way to use an image included with `include_bytes!` without
needing to copy it to a vec `Vec` to create an image handle.

It would be nice if these methods could also be `const`, but that isn't
possible due to the hashing being done.

This is technically a breaking change since `Handle::data()` is public.
But if that is used, it's most likely in used somewhere that only relies
on the type derefing to `&[u8]`.